+++ agent/event_test.go
@@ -226,7 +226,7 @@ func Test_prepareEvent(t *testing.T) {
 			wantErr: true,
 		},
 		{
-			name: "missing check",
+			name: "missing check and metrics",
 			args: args{
 				agent: &Agent{},
 				event: &corev2.Event{
@@ -234,6 +234,40 @@ func Test_prepareEvent(t *testing.T) {
 				},
 			},
 			wantErr: true,
+		},
+		{
+			name: "check event",
+			args: args{
+				agent: &Agent{
+					config: &Config{
+						AgentName: "agent1",
+						Namespace: "default",
+					},
+				},
+				event: &corev2.Event{
+					ObjectMeta: corev2.ObjectMeta{Namespace: "default"},
+					Check:      corev2.FixtureCheck("check1"),
+				},
+			},
+			wantErr:       false,
+			wantNamespace: "default",
+		},
+		{
+			name: "metrics event",
+			args: args{
+				agent: &Agent{
+					config: &Config{
+						AgentName: "agent1",
+						Namespace: "default",
+					},
+				},
+				event: &corev2.Event{
+					ObjectMeta: corev2.ObjectMeta{Namespace: "default"},
+					Metrics:    corev2.FixtureMetrics(),
+				},
+			},
+			wantErr:       false,
+			wantNamespace: "default",
 		},
 		{
 			name: "invalid check",